+++ cypress/e2e/smoke.cy.ts
@@ -1,4 +1,4 @@
-import faker from "@faker-js/faker";
+import { faker } from "@faker-js/faker";
 
 describe("smoke tests", () => {
   it("should allow you to register and login", () => {
