+++ package.json
@@ -54,11 +54,11 @@
     "@vitejs/plugin-react": "^1.3.2",
     "c8": "^7.11.3",
     "cross-env": "^7.0.3",
-    "cypress": "^9.7.0",
+    "cypress": "^10.0.3",
     "esbuild": "^0.14.43",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
-    "happy-dom": "^2.55.0",
+    "happy-dom": "^5.2.0",
     "mock-aws-s3": "^4.0.2",
     "msw": "^0.42.1",
     "npm-run-all": "^4.1.5",
